:lipstick: :: [#342] Collapse 로직 추가
